requirements.txt
annotated-types==0.7.0
anyio==4.11.0
argostranslate==1.10.0
asttokens==3.0.0
beautifulsoup4==4.14.2
blis==1.3.0
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.3
ctranslate2==4.6.0
curl_cffi==0.13.0
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.17
decorator==5.2.1
emoji==2.15.0
executing==2.2.1
filelock==3.20.0
fonttools==4.60.1
frozendict==2.4.6
fsspec==2025.10.0
google-auth==2.41.1
google-genai==1.46.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
libretranslatepy==2.1.1
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multitasking==0.0.12
murmurhash==1.0.13
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
parso==0.8.5
pdfminer.six==20250506
pdfplumber==0.11.7
peewee==3.18.3
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
playwright==1.55.0
pluggy==1.6.0
preshed==3.0.10
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydotenv==0.0.7
pyee==13.0.0
Pygments==2.19.2
PyMuPDF==1.26.5
pyparsing==3.2.5
PyPDF2==3.0.1
pypdfium2==5.0.0
pytest==8.4.2
pytest-base-url==2.1.0
pytest-playwright==0.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
reportlab==4.4.5
requests==2.32.5
rich==14.2.0
rsa==4.9.1
sacremoses==0.1.1
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.4.4
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
stanza==1.10.1
sympy==1.14.0
tenacity==9.1.2
text-unidecode==1.3
thinc==8.3.7
threadpoolctl==3.6.0
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
translate==3.8.0
triton==3.5.0
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.2
websockets==15.0.1
wrapt==2.0.0
xx_sent_ud_sm @ https://github.com/explosion/spacy-models/releases/download/xx_sent_ud_sm-3.8.0/xx_sent_ud_sm-3.8.0-py3-none-any.whl#sha256=1b5fb9321723b395f76458d2a13b7a3673c537eb58c0a9df3778f4e0950180b4
yfinance==0.2.66
